+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [4.1.0-alpha.4](https://github.com/ar-io/ar-io-sdk/compare/v4.1.0-alpha.3...v4.1.0-alpha.4) (2026-07-02)
+
+
+### Bug Fixes
+
+* **arns:** resolve txId from ANT record in resolveArNSName ([ba061cf](https://github.com/ar-io/ar-io-sdk/commit/ba061cf51c74ddf827a678f447193316f76ffb03))
+
+# [4.1.0-alpha.3](https://github.com/ar-io/ar-io-sdk/compare/v4.1.0-alpha.2...v4.1.0-alpha.3) (2026-06-30)
+
+
+### Bug Fixes
+
+* **solana:** drop stale getProgramAccounts ghosts before close_observations batch ([f95937c](https://github.com/ar-io/ar-io-sdk/commit/f95937c5f5e3012153b3c8f0b16adacb720fac9b))
+
+# [4.1.0-alpha.2](https://github.com/ar-io/ar-io-sdk/compare/v4.1.0-alpha.1...v4.1.0-alpha.2) (2026-06-25)
+
+
+### Bug Fixes
+
+* **solana:** getVaults returns numeric vaultId so release/revoke work ([eae22cc](https://github.com/ar-io/ar-io-sdk/commit/eae22cc98297a65237068aec334af3bdfb402238))
+
 # [4.1.0-alpha.1](https://github.com/ar-io/ar-io-sdk/compare/v4.0.3-alpha.1...v4.1.0-alpha.1) (2026-06-22)